Support for accuweather forecast in addition to USA National Weather Service.
By default, the script now "falls back" to accuweather forecast data if the
USA National Weather Service data appears not to be there (based on the
'updatetime' field.

The accuweather plugin uses a different nomenclature, and as such, the
LOCATION input can now have a '|' in it, and the | left hand side is for the
National Weather Service and the right hand side is for the accuweather
location.

Note finally that you can get *just* the Accuweather data by putting
'FORCE_ACCUWEATHER' as the left hand side of the | symbol.